+++ ui/app/components/secret-engine/configure-create-edit.hbs
@@ -21,7 +21,6 @@
       Access to AWS
     </h2>
   {{/if}}
-  {{! Form Fields for all WIF engines }}
   <div class="box is-fullwidth is-sideless">
     {{#if this.version.isEnterprise}}
       {{! Only enterprise users who are configuring wif engines can select a specific access type (account vs wif). }}
